requirements.txt
pdfplumber
python-docx
sentence-transformers
tqdm
scikit-learn
Flask